+++ assets/SwitchPage.4c0d3248.js
@@ -1,0 +1,42 @@
+var m=Object.defineProperty;var r=(e,a)=>m(e,"name",{value:a,configurable:!0});import{S as n}from"./Switch.99c0cd63.js";import{C as h}from"./ComponentInfo.71300db8.js";import{F as p}from"./FormControlLabel.6d9dd0f8.js";import{d as t,a6 as l,m as i,t as f,a as b}from"./index.b0b0be32.js";import"./SwitchBase.39df38f2.js";import"./useFormControl.16e9d810.js";import"./useControlled.017201cb.js";import"./Link.7c23a529.js";import"./copyText.aaeb6752.js";import"./PaperCode.03817e69.js";import"./formControlState.798b6d8d.js";function c(){return t(p,{disabled:!0,get control(){return t(n,{defaultChecked:!0})},label:"Label"})}r(c,"Label");c.code=`import { FormControlLabel, Switch } from "@suid/material";
+
+export default function Label() {
+  return (
+    <FormControlLabel
+      disabled
+      control={<Switch defaultChecked />}
+      label="Label"
+    />
+  );
+}
+`;const S=f("<div></div>"),o={inputProps:{"aria-label":"Switch demo"}};function d(){return(()=>{const e=S.cloneNode(!0);return l(e,t(n,i(o,{defaultChecked:!0})),null),l(e,t(n,o),null),l(e,t(n,i(o,{disabled:!0,defaultChecked:!0})),null),l(e,t(n,i(o,{disabled:!0})),null),e})()}r(d,"BasicSwitches");d.code=`import { Switch } from "@suid/material";
+
+const label = { inputProps: { "aria-label": "Switch demo" } };
+
+export default function BasicSwitches() {
+  return (
+    <div>
+      <Switch {...label} defaultChecked />
+      <Switch {...label} />
+      <Switch {...label} disabled defaultChecked />
+      <Switch {...label} disabled />
+    </div>
+  );
+}
+`;function u(){const[e,a]=b(!1);return t(n,{get checked(){return e()},onChange:(C,s)=>{a(s)},inputProps:{"aria-label":"controlled"}})}r(u,"ControlledSwitch");u.code=`import { Switch } from "@suid/material";
+import { createSignal } from "solid-js";
+
+export default function ControlledSwitch() {
+  const [checked, setChecked] = createSignal(false);
+
+  return (
+    <Switch
+      checked={checked()}
+      onChange={(event, value) => {
+        setChecked(value);
+      }}
+      inputProps={{ "aria-label": "controlled" }}
+    />
+  );
+}
+`;function j(){return t(h,{get name(){return n.name},docsName:"switches",examples:[d,u,c]})}r(j,"SwitchPage");export{j as default};